web/database/api/requirements.txt
alembic==1.4.2
amqp==2.5.2
appdirs==1.4.3
attrs==19.3.0
billiard==3.6.3.0
black==19.10b0
celery==4.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Celery-Helper==1.1.0
Flask-Migrate==2.5.3
flask-redis==0.4.0
Flask-SQLAlchemy==2.4.1
flower==0.9.4
humanize==0.5.1
idna==2.10
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.8
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
mod-wsgi==4.7.1
numpy==1.18.4
pandas==1.0.3
pathspec==0.8.0
psycopg2==2.8.5
pycodestyle==2.5.0
pydocstyle==5.0.2
pyflakes==2.2.0
pylama==7.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
redis==3.5.1
regex==2020.5.7
requests==2.24.0
six==1.14.0
snowballstemmer==2.0.0
SQLAlchemy==1.3.16
toml==0.10.0
tornado==6.0.4
typed-ast==1.4.1
urllib3==1.25.9
vine==1.3.0
Werkzeug==1.0.1
xlrd==1.2.0
zipp==3.1.0
